new stats

+++ packetfilter.h
@@ -32,7 +32,7 @@ public:
 
 private:
     static long _tcp, _udp, _icmp, _igmp, _other;
-    std::map<Poco::Int32, Poco::Int64> _packetIP;
+    std::map<Poco::Int64, Poco::Int64> _packetIP;
     std::map<Poco::Int32, Poco::Int64> _packetPort;
 };
